+++ crates/macros/src/config/field_value.rs
@@ -108,20 +108,20 @@ impl<'l> FieldValue<'l> {
         match self {
             Self::NestedList { .. } => Some(quote! {
                 for (i, item) in setting.iter().enumerate() {
-                    if let Err(nested_error) = item.validate_with_path(context, path.join_key(#name_quoted).join_index(i)) {
+                    if let Err(nested_error) = item.validate_with_path(context, finalize, path.join_key(#name_quoted).join_index(i)) {
                         errors.push(schematic::ValidateErrorType::nested(nested_error));
                     }
                 }
             }),
             Self::NestedMap { .. } => Some(quote! {
                 for (key, value) in setting {
-                    if let Err(nested_error) = value.validate_with_path(context, path.join_key(#name_quoted).join_key(key)) {
+                    if let Err(nested_error) = value.validate_with_path(context, finalize, path.join_key(#name_quoted).join_key(key)) {
                         errors.push(schematic::ValidateErrorType::nested(nested_error));
                     }
                 }
             }),
             Self::NestedValue { .. } => Some(quote! {
-                if let Err(nested_error) = setting.validate_with_path(context, path.join_key(#name_quoted)) {
+                if let Err(nested_error) = setting.validate_with_path(context, finalize, path.join_key(#name_quoted)) {
                     errors.push(schematic::ValidateErrorType::nested(nested_error));
                 }
             }),